Correctly handle carriage returns in messages
In the katcp spec, an <eol> is either \r or \n, but only \n was handled.
The Message class is fixed to treat \r and \n interchangeably.

Because the readuntil API in StreamReader is limited to a single
separator, a protocol wrapper is used to convert \r -> \n as data is
buffered. This is a bit dirty, but the alternative would require
reimplementing most of StreamReader.