Tweak type annotations and add assertions for mypy
This gets it a little closer to passing with --strict-optional, although
there is still more work to be done. It's mostly in places where class
attributes are either initialised later (e.g. in setUp for tests) or set
to None in a close function, and hence have to be Optional but are
guaranteed to be set in most places they're used.