Simplify SensorWatcher a bit
It no longer accepts an external list of sensors. It is easy for a
subclass to mirror into an external list. The previous version had an
issue where a disconnect would mark *all* the sensors in the external
set as UNREACHABLE, rather than just the ones belonging to the client.