Switch from nose to pytest
The tests are moved to a top-level tests directory and the source is
moved down to a src/ directory, to deal with known issues in collecting
coverage with pytest when source is directly off the repo root.

The tests still use unittest style classes rather than taking advantage
of pytest.