Fix ResourceWarnings
For some reason the change to the implementation of DeviceServer.stop
has exposed some errors about streams not being fully closed. I think it
is most likely just a timing issue about how many event loop iterations
are needed.

To fix it, we need to use `writer.wait_closed` in Connection. The
previous Connection design set the `writer` to None once it was asked to
close, but that meant it wasn't available to wait on. That's been
replaced by use of a boolean flag.