Use bulk sensor sampling in SensorMonitor
This is done opportunistically: if it fails, or if the server doesn't
support it, fall back to subscribing to individual sensors. This may be
needed in the case that sensors are removed before we had a chance to
subscribe to them.

This doubles the performance of SensorMonitor in syncing to a server
with a large number of sensors.