Tidy up AggregateSensor
Add a __del__() method which detaches and unregisters things, which
should hopefully make garbage collection better.

Also prepare the way for a new style of "change-aware" observer
functions, which will understand both a new and an old reading, enabling
more efficient aggregation of certain kinds of sensor values such as
totals or counts for instance.

I haven't implemented the new style of change-aware observer functions
yet, but I have an idea of how I'm going to do it. This is just mostly
working and a logical place (in my mind) to commit.