Apply some more PR feedback
Document the filter_aggregate function in the server tutorial.

Move most of the actual code over to the examples directory, with just
the general idea in the actual rst file. This included adding a service
task that added and removed a fixed-value int sensor to show how the
total kept track.

Also change "change_aware" to "delta" in the test file, which I hadn't
done previously.

Also add a new test that the value of the aggregate is unchanged when
the update_aggregate returns None.