Improve signature checking of handlers
wrap_handler attempts to detect and provide a useful error when the
provided arguments are incompatible with the function signature, by
catching the TypeError that ensues. There are two issues with this
however:

1. If the handler is synchronous, the TypeError might come from inside
   the implementation of the handler, rather than the immediate call.
2. If the handler is asynchronous, it only returns the coroutine object
   without directly running it. That solves the first problem, but if
   the handler is decorated with a generic decorator, the TypeError
   might only occur once the decorator tries to run the actual handler,
   and this might occur too late.

Fix both cases by using `inspect.bind` to validate the arguments. While
I'm at it, refactor some of the code to eliminate the rather hacky
`types.coroutine` wrapping that was used to turn a synchronous function
returning a coroutine object into a coroutine function.