Restore time_sync module on macOS
MacOS does not have the `adjtimex` system call. As a result we have
stripped out the entire `time_sync` module on non-Linux platforms.

However, packages like katsdpcontroller need the `ClockState` enum
without the rest. Instead of spreading the complications, restore
the `time_sync` module with a fake `adjtimex` instead. It returns
the correct time but with artificially large errors in the absence
of real values (and the status is always "OK").