Make Reading a data class
Amongst other advantages (such as repr), this makes it implement
`__eq__`, allowing it to be compared in unit tests.

While a reading ought to typically be immutable, I didn't make it frozen
since I don't want to break backwards compatibility for any code that
does modify a reading.

See NGC-1063.