Add a safeguard to sensor_reading helper
aiokatcp and katcp-python support an extension to ?sensor-value that
allows a regex to be passed (between slashes) and returns all matching
sensors. If such a regex is passed to `sensor_reading`, the return would
not contain exactly one sensor value as expected. Add a check that the
return contains exactly one inform.