Fix hypothesis testing of sensor replacements
There is a corner cases of server behaviour that it's impossible to
handle sensibly: if a sensor is replaced with an identical one, we will
no longer be subscribed but cannot detect this.

The hypothesis test skipped directly making such a replacement, but
could still do so by making two replacements that change a property then
change it back. Update the test so that it appends a counter value to
the units every time, so that no replacement ever matches the original.

Relates to NGC-1660