Simplify and improve connection.py
Eliminate _closing and _writer_closing, and just rely on
writer.is_closing(). This reduces the complexity of the state machine.

Avoid skipping `drain` when the connection is closing. That's a
dangerous thing to do, since it removes any backpressure on a closing
connection and allows the buffer to be filled faster than it can be
emptied.

Avoid suppressing errors from `drain`. I'll have to see how that works
out once I finish modifying the rest of the code. It seems like a good
idea for writers to throw when the client has disconnected though.

Catch BrokenPipe error where ConnectionResetError was already being
caught. This prevents annoying error messages appearing in logs when a
client disconnects and the server tries to send it a sensor update for
example.

This commit doesn't pass unit tests. It probably needs some rework of
the Client class to properly implement the state machine.