Avoid using _SelectorSocketTransport.writelines
It causes hangs in the hypothesis tests due to
https://github.com/python/cpython/issues/136234. Specifically, there is
a window between closing the Transport and getting the connection_lost
callback. If called during that window, it will re-add the socket to the
loop for writes, but the actual attempts to send data will always be
skipped, leading to an infinite loop since the file descriptor is always
writable but we never write to it nor remove it from our set of
interest.