Extend hypothesis tests for Client
This revealed the issue that a @precondition is not guaranteed to still
hold by the time the code wrapped by @run_in_event_loop actually runs.
In Python 3.12 that can probably be fixed using an eager task factory,
but that's not an option for 3.8. For now I've worked around it by
re-checking the precondition inside the coroutine, but it does cause
some warnings about `assume` being deprecated outside of property-based
tests.